+++ modules/mod_coalawebflair/tmpl/default.php
@@ -1,27 +1,24 @@
 <?php
-defined('_JEXEC') or die('Restricted access');
+
 /**
- * @package             Joomla
- * @subpackage          CoalaWeb Flair Module
- * @author              Steven Palmer
- * @author url          https://coalaweb.com
- * @author email        support@coalaweb.com
- * @license             GNU/GPL, see /assets/en-GB.license.txt
- * @copyright           Copyright (c) 2017 Steven Palmer All rights reserved.
+ * @package     Joomla
+ * @subpackage  CoalaWeb Flair
+ * @author      Steven Palmer <support@coalaweb.com>
+ * @link        https://coalaweb.com/
+ * @license     GNU/GPL V3 or later; https://www.gnu.org/licenses/gpl-3.0.html
+ * @copyright   Copyright (c) 2021 Steven Palmer All rights reserved.
  *
- * Coala Web Flair is free software: you can redistribute it and/or modify
+ * CoalaWeb Flair is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
-
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
-
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
+
+defined('_JEXEC') or die('Restricted access');
 ?>
 <div class="cw-flair-mod" id="cw-flair-<?php echo $uniqueId; ?>">
     <?php if ($layout) : ?>